requirements.txt
hydra-core==1.1.1
matplotlib==3.5.0
numpy==1.21.2
ogb==1.3.3
omegaconf==2.1.1
pandas==1.4.0
scikit_learn==1.0.2
setuptools==58.0.4
torch==1.10.2
torch_geometric==2.0.3
